Change Packaging Files, added prereqs
Removed setup.cfg and setup.py to conform with modern packaging practices. Updated pyproject.toml and included openai and together as dependencies. May move to make all optional dependencies as we discussed making the client instantiation a prereq to using our library instead of handling it ourselves. Ticked version to 0.0.3.